+++ ebs/src/modules/config/index.ts
@@ -1,16 +1,21 @@
 import { Config } from "common/types";
-import { app } from "..";
-import { sendPubSubMessage } from "../util/pubsub";
+import { sendPubSubMessage } from "../../util/pubsub";
 import { compressSync, strFromU8, strToU8 } from "fflate";
-import { asyncCatch } from "../util/middleware";
-import { Webhooks } from "@octokit/webhooks";
-import { sendToLogger } from "../util/logger";
+import { sendToLogger } from "../../util/logger";
 
-let activeConfig: Config | undefined;
 let configData: Config | undefined;
+let activeConfig: Config | undefined;
+let ingameState = false;
 
 const apiURL = "https://api.github.com/repos/vedalai/swarm-control/contents/config.json";
 const rawURL = "https://raw.githubusercontent.com/VedalAI/swarm-control/main/config.json";
+
+require("./endpoints");
+
+(async () => {
+    const config = await getConfig();
+    await broadcastConfigRefresh(config);
+})().then();
 
 async function fetchConfig(): Promise<Config> {
     let url = `${apiURL}?${Date.now()}`;
@@ -26,17 +31,17 @@ async function fetchConfig(): Promise<Config> {
         console.error("Error when fetching config from api URL, falling back to raw URL");
         console.error(e);
 
-        sendToLogger({
-            transactionToken: null,
-            userIdInsecure: null,
-            important: true,
-            fields: [
-                {
-                    header: "Error when fetching config from api URL, falling back to raw URL",
-                    content: e.toString(),
-                },
-            ],
-        }).then();
+            sendToLogger({
+                transactionToken: null,
+                userIdInsecure: null,
+                important: true,
+                fields: [
+                    {
+                        header: "Error when fetching config from api URL, falling back to raw URL",
+                        content: e.toString(),
+                    },
+                ],
+            }).then();
 
         try {
             url = `${rawURL}?${Date.now()}`;
@@ -47,7 +52,7 @@ async function fetchConfig(): Promise<Config> {
         } catch (e: any) {
             console.error("Error when fetching config from raw URL, panic");
             console.error(e);
-    
+
             sendToLogger({
                 transactionToken: null,
                 userIdInsecure: null,
@@ -66,7 +71,16 @@ async function fetchConfig(): Promise<Config> {
             };
         }
     }
+}
 
+export function isIngame() {
+    return ingameState;
+}
+
+export function setIngame(newIngame: boolean) {
+    if (ingameState == newIngame) return;
+    ingameState = newIngame;
+    setActiveConfig(configData!).then();
 }
 
 function processConfig(data: Config) {
@@ -85,6 +99,14 @@ export async function getConfig(): Promise<Config> {
     return activeConfig!;
 }
 
+export async function getRawConfigData(): Promise<Config> {
+    if (!configData) {
+        await refreshConfig();
+    }
+
+    return configData!;
+}
+
 export async function setActiveConfig(data: Config) {
     activeConfig = processConfig(data);
     await broadcastConfigRefresh(activeConfig);
@@ -97,74 +119,13 @@ export async function broadcastConfigRefresh(config: Config) {
     });
 }
 
-let ingameState: boolean = false;
-
-export function isIngame() {
-    return ingameState;
-}
-
-export function setIngame(newIngame: boolean) {
-    if (ingameState == newIngame) return;
-    ingameState = newIngame;
-    setActiveConfig(configData!).then();
-}
-
 async function refreshConfig() {
     configData = await fetchConfig();
     activeConfig = processConfig(configData);
 }
 
-app.get(
-    "/private/refresh",
-    asyncCatch(async (_, res) => {
-        sendRefresh();
-
-        res.send(configData);
-    })
-);
-
-const webhooks = new Webhooks({
-    secret: process.env.PRIVATE_API_KEY!,
-});
-
-app.post(
-    "/webhook/refresh",
-    asyncCatch(async (req, res) => {
-        // github webhook
-        const signature = req.headers["x-hub-signature-256"] as string;
-        const body = JSON.stringify(req.body);
-
-        if (!(await webhooks.verify(body, signature))) {
-            res.sendStatus(403);
-            return;
-        }
-
-        // only refresh if the config.json file was changed
-        if (req.body.commits.some((commit: any) => commit.modified.includes("config.json"))) {
-            sendRefresh();
-
-            res.status(200).send("Config refreshed.");
-        } else {
-            res.status(200).send("Config not refreshed.");
-        }
-    })
-);
-
-async function sendRefresh() {
+export async function sendRefresh() {
     await refreshConfig();
     console.log("Refreshed config, new config version is ", activeConfig!.version);
     await broadcastConfigRefresh(activeConfig!);
 }
-
-app.get(
-    "/public/config",
-    asyncCatch(async (req, res) => {
-        const config = await getConfig();
-        res.send(JSON.stringify(config));
-    })
-);
-
-(async () => {
-    const config = await getConfig();
-    await broadcastConfigRefresh(config);
-})().then();
