+++ ebs/src/modules/orders/endpoints/index.ts
@@ -1,0 +1,2 @@
+require("./public");
+require("./private");
