+++ ebs/src/modules/orders/endpoints/private.ts
@@ -1,0 +1,6 @@
+import { app } from "../../..";
+import { getOrder } from "../../../util/db";
+
+app.get("/private/order/:guid", (req, res) => {
+    return getOrder(req.params["guid"]);
+});
