+++ ebs/src/modules/twitch/index.ts
@@ -1,5 +1,5 @@
-import { getHelixUser } from "../util/twitch";
-import { TwitchUser } from "./game/messages";
+import { getHelixUser } from "../../util/twitch";
+import { TwitchUser } from "../game/messages";
 
 export async function getTwitchUser(id: string): Promise<TwitchUser | null> {
     const user = await getHelixUser(id);
